fix some old migrations

+++ db/migrate/20170315182905_remove_user_widgets.rb
@@ -1,5 +1,5 @@
 class RemoveUserWidgets < ActiveRecord::Migration[5.0]
   def change
-    drop_table :user_widgets
+    drop_table :user_widgets if table_exists? :user_widgets
   end
 end
